claims_extraction/requirements.txt
requests==2.31.0
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
transformers==4.35.2